# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ServiceTrack Tesla API
  version: 0.1.0
  description: API для однієї незалежної майстерні Tesla. Сесійна автентифікація та
    CSRF.
paths:
  /api/customers/:
    get:
      operationId: customers_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - customers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomerList'
              example:
                count: 1
                next: null
                previous: null
                results:
                - id: 1
                  name: Демо клієнт
                  phone: '+380000000001'
                  email: demo@example.invalid
                  notes: ''
                  created_at: '2026-09-29T13:33:27.005687+03:00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
    post:
      operationId: customers_create
      tags:
      - customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
            example:
              name: Демо клієнт
              phone: '+380000000001'
              email: demo@example.invalid
              notes: ''
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Customer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: 1
                name: Демо клієнт
                phone: '+380000000001'
                email: demo@example.invalid
                notes: ''
                created_at: '2026-09-29T13:33:27.005687+03:00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '415':
          description: Непідтримуваний Content-Type.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      parameters:
      - in: header
        name: X-CSRFToken
        required: true
        description: Актуальний CSRF після входу; cookie sessionid і csrftoken надсилаються
          разом.
        schema:
          type: string
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
  /api/customers/{id}/:
    get:
      operationId: customers_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      tags:
      - customers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: 1
                name: Демо клієнт
                phone: '+380000000001'
                email: demo@example.invalid
                notes: ''
                created_at: '2026-09-29T13:33:27.005687+03:00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
    put:
      operationId: customers_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      - in: header
        name: X-CSRFToken
        required: true
        description: Актуальний CSRF після входу; cookie sessionid і csrftoken надсилаються
          разом.
        schema:
          type: string
      tags:
      - customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
            example:
              name: Демо клієнт
              phone: '+380000000001'
              email: demo@example.invalid
              notes: ''
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Customer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: 1
                name: Демо клієнт
                phone: '+380000000001'
                email: demo@example.invalid
                notes: ''
                created_at: '2026-09-29T13:33:27.005687+03:00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '415':
          description: Непідтримуваний Content-Type.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
    patch:
      operationId: customers_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      - in: header
        name: X-CSRFToken
        required: true
        description: Актуальний CSRF після входу; cookie sessionid і csrftoken надсилаються
          разом.
        schema:
          type: string
      tags:
      - customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
            example:
              notes: Уточнення
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCustomer'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              example:
                id: 1
                name: Демо клієнт
                phone: '+380000000001'
                email: demo@example.invalid
                notes: Уточнення
                created_at: '2026-09-29T13:33:27.005687+03:00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '415':
          description: Непідтримуваний Content-Type.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
    delete:
      operationId: customers_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this customer.
        required: true
      - in: header
        name: X-CSRFToken
        required: true
        description: Актуальний CSRF після входу; cookie sessionid і csrftoken надсилаються
          разом.
        schema:
          type: string
      tags:
      - customers
      security:
      - cookieAuth: []
      responses:
        '204':
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
  /api/items/:
    get:
      operationId: items_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - items
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedItemList'
              example:
                count: 1
                next: null
                previous: null
                results:
                - id: 1
                  order: 1
                  kind: labor
                  name: Огляд
                  quantity: '1.50'
                  unit_price: '1000.00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
    post:
      operationId: items_create
      tags:
      - items
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
            example:
              order: 1
              kind: labor
              name: Огляд
              quantity: '1.50'
              unit_price: '1000.00'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Item'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Item'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              example:
                id: 1
                order: 1
                kind: labor
                name: Огляд
                quantity: '1.50'
                unit_price: '1000.00'
          description: Успішна відповідь, приклад із ізольованої демонстраційної БД.
        '400':
          description: Некоректні поля, фільтр або бізнес-правило.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '403':
          description: Немає сесії, прав або коректного CSRF.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '404':
          description: Об’єкт або сторінка не знайдені / недоступні.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '405':
          description: Метод не підтримується.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '406':
          description: Непідтримуваний Accept.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '415':
          description: Непідтримуваний Content-Type.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '429':
          description: Перевищено 600 запитів за хвилину для користувача; Retry-After.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                description: Помилки полів або detail; текст локалізований.
        '500':
          description: Неочікувана помилка сервера; формат залежить від DEBUG і сервера.
      parameters:
      - in: header
        name: X-CSRFToken
        required: true
        description: Актуальний CSRF після входу; cookie sessionid і csrftoken надсилаються
          разом.
        schema:
          type: string
      description: Потрібна сесія. Механік читає лише призначені йому замовлення та
        пов’язані записи. Зміни customers/vehicles/items доступні лише адміністратору.
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 30
        email:
          type: string
          format: email
          maxLength: 254
        notes:
          type: string
          maxLength: 4000
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
      - phone
...
    Item:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        order:
          type: integer
        kind:
          $ref: '#/components/schemas/ItemKindEnum'
        name:
          type: string
          maxLength: 200
        quantity:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,2})?$
        unit_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
      required:
      - id
      - kind
      - name
      - order
      - quantity
      - unit_price
    ItemKindEnum:
      enum:
      - labor
      - part
      type: string
      description: |-
        * `labor` - Робота
        * `part` - Запчастина
...
    PaginatedCustomerList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
...
    PaginatedItemList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Item'
...
    PatchedCustomer:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 120
        phone:
          type: string
          maxLength: 30
        email:
          type: string
          format: email
          maxLength: 254
        notes:
          type: string
          maxLength: 4000
        created_at:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid